Lume_project/requirements.txt
discord.py==2.6.0a5260+g4be15b56
discord-ext-voice-recv==0.5.2a179
faster-whisper==1.1.1
APScheduler==3.11.0
#torch torchvision torchaudio --index-url https://download.pytorch.org/whl/cu128
fuzzywuzzy==0.18.0
NRCLex==3.0.0
dotenv
sounddevice==0.5.2
soundfile==0.13.1
Levenshtein==0.27.1
librosa==0.9.2
scipy==1.15.3
gradio==5.29.1
transformers==4.51.3
peft==0.15.2
PyQt5==5.15.11
jieba==0.42.1
fast-langdetect==0.3.2
split-lang==2.1.0
x-transformers==2.3.5
PyNaCl==1.5.0
pytorch-lightning==2.5.1.post0
matplotlib==3.10.3
cn2an==0.5.23
pypinyin==0.54.0
pyvts==0.3.3 #live2d
onnxruntime==1.22.0
onnxruntime-gpu==1.22.0
standard-aifc==3.13.0
standard-sunau==3.13.0
standard-chunk==3.13.0
wordsegment==1.3.1
g2p-en==2.1.0
opencv-python-headless 
easyocr==1.7.2
pytesseract==0.3.13
pyautogui==0.9.54
mss==10.0.0
sentence-transformers>=2.2.0#memory
chromadb#memory
flask[async] #api
waitress#api
twitchapi #twitch
#jieba_fast